+++ src/lime/protocol/envelope.py
@@ -11,15 +11,3 @@ class Envelope(ABC, Serializable):
         self.to = to
         self.pp = pp
         self.metadata = metadata
-
-
-class EnvelopeListener(ABC):
-    """Envelope listener callbacks."""
-
-    def on_envelope(self, envelope: Envelope):
-        """Handle callback to envelope received event.
-
-        Args:
-            envelope (Envelope): the received envelope
-        """
-        pass
